created reward system

+++ client/src/actions/question.js
@@ -40,6 +40,7 @@ export const postAnswer=(answerdata)=>async(dispatch)=>{
     const {id, noOfAnswers, answerBody, userAnswered, userId}=answerdata;
 
   const {data}=await api.postAnswer(id, noOfAnswers, answerBody, userAnswered, userId);
+
   console.log('userid of post answer is :'+JSON.stringify(answerdata));
   dispatch({type:'POST_ANSWER', payload:data});
 dispatch(fetchAllQuestions());
@@ -48,10 +49,10 @@ dispatch(fetchAllQuestions());
   }
 }
 
-export const deleteAnswer=(id, answerId, noOfAnswers)=>async(dispatch)=>{
+export const deleteAnswer=(id, answerId, noOfAnswers,userId)=>async(dispatch)=>{
   try{
     console.log("right id is"+id);
-const {data}=await api.deleteAnswer(id, answerId, noOfAnswers);
+const {data}=await api.deleteAnswer(id, answerId, noOfAnswers,userId);
 dispatch(fetchAllQuestions());
   }catch(error){
 console.log(error);
@@ -68,4 +69,15 @@ try{
 console.log(error);
 }
 
+}
+
+export const countUpvotedQuestions=(userId)=>async(dispatch)=>{
+try{
+console.log("countupvotd clicked"+userId);
+let {data}=await api.countUpvotedQuestions(userId);
+data=data.data;
+dispatch({type:"setQuestionCountScore", payload:data});
+}catch(error){
+console.log(error);
+}
 }
